backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic-settings==2.5.2

# GPX Processing
gpxpy==1.6.2

# Authentication
authlib==1.3.2
httpx==0.27.2
itsdangerous==2.2.0

# Google APIs (Phase 2)
# google-api-python-client==2.149.0
# google-auth-httplib2==0.2.0
# google-auth-oauthlib==1.2.1

# CORS
fastapi-cors==0.0.6

# Environment
python-dotenv==1.0.1

# Database (Phase 3 - Sharing)
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Processing
pandas==2.2.3
numpy==2.1.2
